+++ packages/sdk-ink/src/sdk-types.ts
@@ -54,6 +54,21 @@ export interface ReviveSdk<
   addressIsMapped: (address: SS58String) => Promise<boolean>
 }
 
+export interface InkSdkOptions {
+  /**
+   * Target the latest block instead of the finalized block for any query or dry-run operation.
+   *
+   * This makes it possible to have quicker updates, but be mindful that the data returned might become
+   * invalid at any point, e.g. a contract that apparently was successfully deployed might suddenly
+   * disappear, just to reappear a few seconds later, or maybe never? Be really mindful
+   * you can get inconsistencies.
+   */
+  atBest: boolean
+}
+export const defaultOptions: InkSdkOptions = {
+  atBest: false,
+}
+
 type DryRunDeployFn<
   T extends InkSdkTypedApi | ReviveSdkTypedApi,
   Addr,